Add interface mpas_dmpar_compact_halo_info
This new interface currently has two implementations: for 2-d and 3-d real
fields. The purpose of these routines is to take the information needed to
exchange a field's halo and to compact it into a simple, compact array of
integer values that can be used in the mpas_dmpar_exch_halo_acc routines.