Removing the 'update device' and 'update host' directives
note: this code contains all the code till PR#14. The acceptance is
dependent on the acceptance of PRs till PR#14

The update device and update host directives are removed to facilitate
the direct GPU-GPU halo exchange. Previously the halo-MPI communication
was happening through the host. So the data was being transferred between
host and device before and after each halo exchange call. This code contains
the newly implemented gpu_halo exhchange calls which are suitable for
GPU-GPU communication.