Compute compactHaloInfo arrays needed by scalar transport in mpas_atm_dynamics_init
The monotonic scalar transport code was computing compactHaloInfo arrays
in each call to the atm_advance_scalars_mono_work routine. One of these arrays
was used to exchange halos for a scalar array, and the other was used to exchange
halos for scaling factors. The former can be re-used from the compactHaloInfo_c_v_s
array, and the latter is now computed in the mpas_atm_dynamics_init routine.

As a result, we no longer need to call mpas_dmpar_compact_halo_info from within
the monotonic scalar transport code.