Removal of IfDef OpenACC and unnecessary comments
Removed the IfDef MPAS_OPENACC to make it work for GPUs.
The code inside ifDef was not executed resulting in the compilation
error. This code is necessary for assigning rank to GPU, when a single
host has multiple GPU in it. And also I have removed the commented
lines which are unnecessary.